+++ src/Resources/contao/config/config.php
@@ -12,7 +12,7 @@
 
 use HeimrichHannot\FieldpaletteBundle\EventListener\Contao\ExecutePostActionsListener;
 use HeimrichHannot\FieldpaletteBundle\EventListener\Contao\InitializeSystemListener;
-use HeimrichHannot\FieldpaletteBundle\EventListener\LoadDataContainerListener;
+use HeimrichHannot\FieldpaletteBundle\Model\FieldPaletteModel;
 use HeimrichHannot\FieldpaletteBundle\Widget\FieldPaletteWizard;
 
 $GLOBALS['TL_CONFIG']['fieldpalette_table'] = 'tl_fieldpalette';
@@ -25,11 +25,10 @@ $GLOBALS['BE_FFL'][FieldPaletteWizard::TYPE] = FieldPaletteWizard::class;
 /*
  * Hooks
  */
-$GLOBALS['TL_HOOKS']['loadDataContainer']['fieldPalette'] = [LoadDataContainerListener::class, 'onLoadDataContainer'];
 $GLOBALS['TL_HOOKS']['initializeSystem']['fieldPalette'] = [InitializeSystemListener::class, '__invoke'];
 $GLOBALS['TL_HOOKS']['executePostActions']['fieldPalette'] = [ExecutePostActionsListener::class, '__invoke'];
 
 /*
  * Models
  */
-$GLOBALS['TL_MODELS']['tl_fieldpalette'] = HeimrichHannot\FieldpaletteBundle\Model\FieldPaletteModel::class;
+$GLOBALS['TL_MODELS']['tl_fieldpalette'] = FieldPaletteModel::class;